+++ src/views/testview.cpp
@@ -27,7 +27,7 @@ TestView::TestView(std::shared_ptr<GLWindow> window)
 
     AddBox(Vec3(0.0f, -200.0f, 0.0f), Vec3(70.0f, 50.0f, 90.0f), Vec3(0.0f, 0.0f, 0.0f));
     AddBox(Vec3(0.0f, -500.0f, 0.0f), Vec3(2000.0f, 10.0f, 2000.0f), Vec3(0.0f, 0.0f, 0.0f));
-    AddSphere(100.0f, Vec3(sinf(m_ObjectLocation) * m_ObjectDistance, -300.0f, cosf(m_ObjectLocation) * m_ObjectDistance), 20, 20);
+    AddSphere(100.0f, Vec3(std::sin(m_ObjectLocation) * m_ObjectDistance, -300.0f, std::cos(m_ObjectLocation) * m_ObjectDistance), 20, 20);
     AddLightSource(LightSource::POINT_LIGHT, Vec3(0.0f, 0.0f, 0.0f), Vec3(1.0f, 1.0f, 1.0f), 1.0f);
 
     InitRendering();
@@ -251,7 +251,7 @@ void TestView::DoTick()
 
     if (m_Spheres.size() >= 1)
     {
-        m_Spheres[0].m_Center = Vec3(sinf(m_ObjectLocation) * m_ObjectDistance, -300.0f, cosf(m_ObjectLocation) * m_ObjectDistance);
+        m_Spheres[0].m_Center = Vec3(std::sin(m_ObjectLocation) * m_ObjectDistance, -300.0f, std::cos(m_ObjectLocation) * m_ObjectDistance);
         m_Spheres[0].m_Rotation.x = 0.0f;        
         m_Spheres[0].m_Rotation.y = m_ObjectRotation;
         m_Spheres[0].m_Rotation.z = 23.5;
